// src/app/modules/cpassapi/model/cpv.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SettoreInterventi } from './settoreInterventi';


export interface Cpv { 
    id?: number;
    idPadre?: number;
    livello?: number;
    codice?: string;
    descrizione?: string;
    codicePadre?: string;
    tipologia?: string;
    divisione?: string;
    gruppo?: string;
    classe?: string;
    categoria?: string;
    settoreInterventi?: SettoreInterventi;
    listCpv?: Array<Cpv>;
}