// src/app/modules/cpassapi/model/intervento.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AcquistoVariato } from './acquistoVariato';
import { Ausa } from './ausa';
import { Cpv } from './cpv';
import { InterventoAltriDati } from './interventoAltriDati';
import { InterventoImporti } from './interventoImporti';
import { MetadatiFunzione } from './metadatiFunzione';
import { ModalitaAffidamento } from './modalitaAffidamento';
import { Nuts } from './nuts';
import { Priorita } from './priorita';
import { Programma } from './programma';
import { RicompresoTipo } from './ricompresoTipo';
import { Settore } from './settore';
import { SettoreInterventi } from './settoreInterventi';
import { Stato } from './stato';
import { Utente } from './utente';


export interface Intervento { 
    id?: string;
    cui?: string;
    annoAvvio?: number;
    cup?: string;
    lottoFunzionale?: boolean;
    esenteCup?: boolean;
    durataMesi?: number;
    nuovoAffidamento?: boolean;
    interventoCopiaTipo?: string;
    interventoImportiCopiaTipo?: string;
    flagCuiNonGenerato?: boolean;
    motivazioneNonRiproposto?: string;
    descrizioneAcquisto?: string;
    ausa?: Ausa;
    utenteRup?: Utente;
    ricompresoCui?: string;
    interventoCopia?: Intervento;
    settoreInterventi?: SettoreInterventi;
    settore?: Settore;
    cpv?: Cpv;
    programma?: Programma;
    nuts?: Nuts;
    priorita?: Priorita;
    modalitaAffidamento?: ModalitaAffidamento;
    stato?: Stato;
    acquistoVariato?: AcquistoVariato;
    ricompresoTipo?: RicompresoTipo;
    listInterventoImporti?: Array<InterventoImporti>;
    listCpv?: Array<Cpv>;
    listInterventoAltriDati?: Array<InterventoAltriDati>;
    listMetadatiFunzione?: Array<MetadatiFunzione>;
    risorsaIdCapitalePrivato?: number;
    dataCreazione?: Date;
    utenteCreazione?: string;
    dataModifica?: Date;
    utenteModifica?: string;
    dataCancellazione?: Date;
    utenteCancellazione?: string;
    optlock?: string;
    dataVisto?: Date;
    dataValidazione?: Date;
    dataRifiuto?: Date;
    utenteVisto?: Utente;
    utenteValidazione?: Utente;
    utenteRifiuto?: Utente;
}